Remove Ctrl-L because it breaks the default (clear/redraw)
Ctrl-L in Vim works with the same principle as in Bash:
to clear the view, which in Vim means "nohighlight" and "redraw".

Overriding it removes an important functionality for those who rely on
vanilla Vim.

However, these dotfiles already provide an unobtrusive setting
(Alt-Right and Alt-Left) which can be used with the same purpose (they
also work seamlessly between Tmux and Vim).