[vim] [shell] Simplify ripgrep configuration
It turns out that $HOME/.rgignore is always loaded by default by ripgrep
command, so there is no need to reference it in the $HOME/.ripgreprc.

This is actually never specified in `man rg`, where it seems to imply
that .rgignore is loaded from the current directory.

It also works well with Vim out of the box, so unnecessary config has
been removed from the vimrc too.